+++ src/scenes/Dashboard/cells/logo.jsx
@@ -1,14 +1,31 @@
 import React from 'react';
-import hackillinoisLogoLightBg from 'assets/logo.svg';
-import hackillinoisLogoDarkBg from 'assets/logo_dark_bg.svg';
+import hackillinoisLogoDay from 'assets/logo.svg';
+import hackillinoisLogoNight from 'assets/logo_dark_bg.svg';
 import exploreConnectBuild from 'assets/home/tagline.svg';
+import ThemeContext from '../theme-context';
 
-function Logo() {
+function renderTheme(theme) {
+  if (theme === 'day') {
+    return (
+      <div className="cell short-cell" id="logo-cell">
+        <img src={exploreConnectBuild} id="explore-connect-build" alt="Explore. Connect. Build." />
+        <img src={hackillinoisLogoDay} id="logo" alt="HackIllinois Logo" />
+      </div>
+    );
+  }
   return (
     <div className="cell short-cell" id="logo-cell">
       <img src={exploreConnectBuild} id="explore-connect-build" alt="Explore. Connect. Build." />
-      <img src={hackillinoisLogoDarkBg} id="logo" alt="HackIllinois Logo" />
+      <img src={hackillinoisLogoNight} id="logo" alt="HackIllinois Logo" />
     </div>
+  );
+}
+
+function Logo() {
+  return (
+    <ThemeContext.Consumer>
+      { value => renderTheme(value)}
+    </ThemeContext.Consumer>
   );
 }
