+++ src/scenes/Dashboard/cells/timeAndEvent.jsx
@@ -1,23 +1,7 @@
 import React from 'react';
 import { getEvents } from 'api';
+import EventBlock from 'scenes/Dashboard/dashboardComponent/eventblock';
 import ThemeContext from '../theme-context';
-
-function EventBlock(props) {
-  const { title, locations, eventTime } = props;
-  return (
-    <div className="event-block">
-      <div className="details">
-        <h2 className="title">{title}</h2>
-        {
-          locations.map(val => <h3 key={`${val}`} className="location">{val}</h3>)
-        }
-      </div>
-      <div className="time">
-        <h2>{eventTime}</h2>
-      </div>
-    </div>
-  );
-}
 
 function pad(num) {
   if (num < 10) {
