+++ src/scenes/Dashboard/styles/_cells.scss
@@ -57,8 +57,6 @@
   }
 }
 
-
-
 #twitter-cell{
   background-color: rgba(0,0,0,0);
   height: calc(100% + 5px)
